backend/requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
torch>=2.1.0
torchvision>=0.16.0
Pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
mlflow==2.13.0
